# specs/001-secure-multitenant-saas/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Multi-Tenant SaaS API
  version: 0.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Health check público
      security: []
      responses:
        '200':
          description: API saudável
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /v1/me:
    get:
      summary: Retorna usuário autenticado e contexto de tenant
      responses:
        '200':
          description: Contexto autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/tenants/current:
    get:
      summary: Retorna tenant atual
      responses:
        '200':
          description: Tenant atual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      summary: Atualiza tenant atual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 120
                slug:
                  type: string
                  pattern: '^[a-z0-9-]{3,60}$'
      responses:
        '200':
          description: Tenant atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/members:
    get:
      summary: Lista membros do tenant atual
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista paginada de membros
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Member'
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/members/{memberId}/role:
    patch:
      summary: Altera papel de membro
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Membro atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/audit-events:
    get:
      summary: Lista eventos de auditoria do tenant atual
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Eventos de auditoria
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /webhooks/clerk:
    post:
      summary: Recebe webhooks Clerk
      security: []
      responses:
        '202':
          description: Evento aceito
        '400':
          description: Assinatura inválida ou payload inválido
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Role:
      type: string
      enum: [owner, admin, member, viewer]
    Tenant:
      type: object
      required: [id, name, slug, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, archived]
    Member:
      type: object
      required: [id, tenantId, clerkUserId, email, role, status]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        clerkUserId:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        status:
          type: string
          enum: [active, invited, disabled]
    AuditEvent:
      type: object
      required: [id, tenantId, action, entityType, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        actorClerkUserId:
          type: string
          nullable: true
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
    MeResponse:
      type: object
      required: [user, tenants]
      properties:
        user:
          type: object
          required: [clerkUserId]
          properties:
            clerkUserId:
              type: string
            email:
              type: string
              nullable: true
        activeTenant:
          $ref: '#/components/schemas/Tenant'
        role:
          $ref: '#/components/schemas/Role'
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
    ErrorResponse:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string